Move all finder keywords to separate class
