+++ packages/main/src/RPA/Desktop/keywords/__init__.py
@@ -1,9 +1,10 @@
 # flake8: noqa
 from robot.api.deco import keyword
-
 from .context import LibraryContext
+
 from .application import ApplicationKeywords
 from .clipboard import ClipboardKeywords
+from .finder import FinderKeywords
 from .keyboard import KeyboardKeywords
 from .mouse import MouseKeywords
 from .screen import ScreenKeywords
